+++ src/tokens.ts
@@ -1,4 +1,6 @@
-import { OpaqueToken } from '@angular/core';
+import { InjectionToken } from '@angular/core';
 
-export const CLIENT_MAP_WRAPPER = new OpaqueToken('apollo/client-map-wrapper');
-export const CLIENT_MAP = new OpaqueToken('apollo/client-map');
+import { ClientMapWrapper, ClientMap } from './types';
+
+export const CLIENT_MAP_WRAPPER = new InjectionToken<ClientMapWrapper>('apollo/client-map-wrapper');
+export const CLIENT_MAP = new InjectionToken<ClientMap>('apollo/client-map');